pyproject.toml
[project]
name = "news-hub-solution-artifacts"
version = "0.1.0"
description = ""
authors = [
    {name = "Lizzy Poker Face",email = "165659964+lizzypokerface@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "google-api-python-client (>=2.181.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "langchain-core (>=0.3.75,<0.4.0)",
    "youtube-transcript-api (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "langchain-ollama (>=0.3.7,<0.4.0)"
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
